Desmembramento de componentes de checkout
