+++ src/Data/dataCupons.js
@@ -1,0 +1,27 @@
+export const Cupons = [
+  {
+    codigo: "cupom10",
+    desconto: 0.1,
+    validade: new Date("2022/12/31 11:59:59 PM"),
+  },
+  {
+    codigo: "cupom15",
+    desconto: 0.15,
+    validade: new Date("2022/12/31 11:59:59 PM"),
+  },
+  {
+    codigo: "cupom25",
+    desconto: 0.25,
+    validade: new Date("2022/12/31 11:59:59 PM"),
+  },
+  {
+    codigo: "cupom30",
+    desconto: 0.3,
+    validade: new Date("2022/12/31 11:59:59 PM"),
+  },
+  {
+    codigo: "cupom50",
+    desconto: 0.5,
+    validade: new Date("2022/12/31 11:59:59 PM"),
+  },
+];
