+++ src/Data/dataProduct.js
@@ -1,74 +1,83 @@
 export const Produtos = [
-    {
-        id: 1,
-        name: "Camisa estrela",
-        price: 29.9,
-        image: "/img/produto1.png",
+  {
+    id: 1,
+    name: "Camisa estrela",
+    price: 29.9,
+    int_price: 0,
+    image: "/img/produto1.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 2,
-        name: "Camisa virgem",
-        price: 69.9,
-        image: "/img/produto2.png",
+    sizeSelect: "",
+  },
+  {
+    id: 2,
+    name: "Camisa virgem",
+    price: 69.9,
+    int_price: 0,
+    image: "/img/produto2.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 3,
-        name: "Camisa aquário",
-        price: 59.9,
-        image: "/img/produto3.png",
+    sizeSelect: "",
+  },
+  {
+    id: 3,
+    name: "Camisa aquário",
+    price: 59.9,
+    int_price: 0,
+    image: "/img/produto3.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 4,
-        name: "Camisa câncer",
-        price: 49.9,
-        image: "/img/produto4.png",
+    sizeSelect: "",
+  },
+  {
+    id: 4,
+    name: "Camisa câncer",
+    price: 49.9,
+    int_price: 0,
+    image: "/img/produto4.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 5,
-        name: "Camisa touro",
-        price: 79.9,
-        image: "/img/produto5.png",
+    sizeSelect: "",
+  },
+  {
+    id: 5,
+    name: "Camisa touro",
+    price: 79.9,
+    int_price: 0,
+    image: "/img/produto5.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 6,
-        name: "Camisa gêmeos",
-        price: 99.9,
-        image: "/img/produto6.png",
+    sizeSelect: "",
+  },
+  {
+    id: 6,
+    name: "Camisa gêmeos",
+    price: 99.9,
+    int_price: 0,
+    image: "/img/produto6.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 7,
-        name: "Camisa leão",
-        price: 59.9,
-        image: "/img/produto7.png",
+    sizeSelect: "",
+  },
+  {
+    id: 7,
+    name: "Camisa leão",
+    price: 59.9,
+    int_price: 0,
+    image: "/img/produto7.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 8,
-        name: "Camisa escorpião",
-        price: 69.9,
-        image: "/img/produto8.png",
+    sizeSelect: "",
+  },
+  {
+    id: 8,
+    name: "Camisa escorpião",
+    price: 69.9,
+    int_price: 0,
+    image: "/img/produto8.png",
 
-        sizeSelect: "",
-    },
-    {
-        id: 9,
-        name: "Camisa sagitário",
-        price: 89.9,
-        image: "/img/produto9.png",
+    sizeSelect: "",
+  },
+  {
+    id: 9,
+    name: "Camisa sagitário",
+    price: 89.9,
+    int_price: 0,
+    image: "/img/produto9.png",
 
-        sizeSelect: "",
-    },
+    sizeSelect: "",
+  },
 ];
