+++ src/Pages/Checkout.css
@@ -1,0 +1,34 @@
+/* Compomente Checkout */
+main {
+  padding: 2rem;
+  background-color: #f4f4f3;
+}
+form {
+  grid-area: main;
+  display: grid;
+  grid-template-areas:
+    "pedido resumo"
+    "envio resumo";
+  grid-template-columns: 1fr 1fr;
+  grid-row-gap: 20px;
+  grid-column-gap: 20px;
+}
+
+@media (max-width: 600px) {
+  form {
+    grid-template-areas:
+      "pedido"
+      "envio"
+      "resumo";
+  }
+}
+
+section {
+  background-color: white;
+  padding: 15px;
+}
+
+td {
+  padding: 4px;
+  border: none;
+}
