+++ src/Routes/routes.jsx
@@ -2,16 +2,16 @@ import { Routes, Route } from "react-router-dom";
 import Home from "../Pages/Home";
 import Product from "../Pages/Product";
 import Login from "../Pages/Login";
-import Checkout from "../components/Checkout/Checkout";
+import Checkout from "../Pages/Checkout";
 
 function ReoutesPages() {
-    return (
-        <Routes>
-            <Route path='/' element={<Home />} />
-            <Route path='/login' element={<Login />} />
-            <Route path='/product/:id' element={<Product />} />
-            <Route path='/checkout' element={<Checkout />} />
-        </Routes>
-    );
+  return (
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/product/:id" element={<Product />} />
+      <Route path="/checkout" element={<Checkout />} />
+    </Routes>
+  );
 }
 export default ReoutesPages;
