--- src/components/Checkout/Checkout.css
@@ -1,305 +1,0 @@
-/* Compomente Checkout */
-main {
-  padding: 2rem;
-  background-color: #f4f4f3;
-}
-form {
-  grid-area: main;
-  display: grid;
-  grid-template-areas:
-    "pedido resumo"
-    "envio resumo";
-  grid-template-columns: 1fr 1fr;
-  grid-row-gap: 20px;
-  grid-column-gap: 20px;
-}
-
-@media (max-width: 600px) {
-  form {
-    grid-template-areas:
-      "pedido"
-      "envio"
-      "resumo";
-  }
-}
-
-section {
-  background-color: white;
-  padding: 15px;
-}
-
-td {
-  padding: 4px;
-  border: none;
-}
-
-/* Compomente Pedido */
-
-#pedido {
-  grid-area: pedido;
-}
-
-#pedido span {
-  display: inline-block;
-}
-
-#pedido span#return {
-  float: right;
-}
-
-#pedido span#return a {
-  font-size: 14px;
-}
-
-#pedido a {
-  font-size: 10px;
-  color: #666;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-#pedido table {
-  width: 98%;
-}
-
-#pedido table td {
-  padding: 12px;
-}
-#pedido table td#preco {
-  width: 120px;
-}
-#pedido table td#empty_list {
-  padding: 0;
-  text-align: center;
-  font-style: italic;
-  line-height: 8.5;
-}
-
-.produto_img {
-  position: relative;
-}
-
-.produto_img img {
-  width: 145px;
-  height: auto;
-}
-
-.produto_nome {
-  font-size: 16px;
-  font-weight: 700;
-}
-
-.produto_ref {
-  font-size: 12px;
-  color: #747474;
-  line-height: 1.85;
-  margin-bottom: 4px;
-}
-
-.produto_atributo {
-  text-align: left;
-  text-decoration: none;
-  font-size: 12px;
-  line-height: 2;
-  /*margin: 8px 0;*/
-}
-
-.delete_produto {
-  text-align: right;
-  margin: 6px;
-}
-
-.produto_preco {
-  text-align: right;
-  font-size: 14px;
-}
-.produto_total {
-  margin-top: 25px;
-  font-weight: bold;
-}
-.dashed {
-  color: #747474;
-  margin-right: 4px;
-  font-size: 12px;
-  text-decoration: line-through;
-}
-.stepper {
-  display: flex;
-  border: 1px solid #e8e8e9;
-  border-radius: var(--radius);
-  background-color: #f8f8f9;
-  height: 32px;
-  width: 88px;
-  margin-top: 6px;
-}
-
-.stepper__button {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  width: 32px;
-  background: none;
-  border: none;
-  font-size: 2.5em;
-  cursor: pointer;
-}
-
-.stepper__content {
-  flex: 1;
-}
-
-.stepper__value {
-  height: 100%;
-  width: 100%;
-  border: none;
-  background: none;
-  text-align: center;
-  padding-top: 4px;
-}
-
-/* Compomente Envio */
-#envio {
-  grid-area: envio;
-}
-
-details {
-  border: 1px solid #ccc;
-  border-radius: var(--radius);
-  margin: 15px 0;
-}
-details summary::-webkit-details-marker {
-  display: none;
-}
-details summary {
-  padding: 15px;
-  cursor: pointer;
-  list-style: none;
-  position: relative;
-  width: 100%;
-  background-color: #efefef;
-  border-radius: inherit;
-}
-details summary::after {
-  content: "+";
-  position: absolute;
-  padding-right: 5px;
-  font-size: 2rem;
-  line-height: 0;
-  margin-top: 0.75rem;
-  margin-right: 12px;
-  right: 0;
-  font-weight: 200;
-  transform-origin: center;
-  transition: 200ms linear;
-}
-
-details[open] summary {
-  border-bottom: 1px solid #ccc;
-}
-
-details[open] summary ~ * {
-  animation: open 0.3s ease-in-out;
-}
-
-details[open] summary:after {
-  transform: rotate(45deg);
-  font-size: 2.25rem;
-}
-
-@keyframes open {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-details summary::-webkit-details-marker {
-  display: none;
-}
-#endereco_cont {
-  padding: 15px;
-}
-#endereco_cont:last-child {
-  border-bottom: none;
-}
-#pagamentos {
-  padding: 15px;
-}
-
-.forma_pagamento_titulo {
-  font-size: 13px;
-  font-weight: bold;
-  padding: 10px 0;
-}
-
-.forma_pagamento_desc {
-  font-size: 10px;
-  color: #747474;
-  text-align: right;
-}
-.forma_pagamento_img {
-  width: 20px;
-  height: auto;
-  border-radius: 5px;
-  margin-right: 10px;
-}
-
-#obs_label {
-  margin: 20px 0;
-}
-
-textarea {
-  width: 100%;
-  padding: 15px;
-  border: 1px solid var(--destaque);
-  border-radius: var(--radius);
-}
-
-/* Compomente Resumo */
-#resumo {
-  grid-area: resumo;
-}
-
-#resumo table {
-  margin: 0 auto;
-  width: 98%;
-  line-height: 1.75;
-}
-
-#resumo table tbody tr .row-div {
-  margin: 7px 0;
-  padding: 3px;
-  border-bottom: 1px solid #efefef;
-}
-
-#resumo table tbody tr td p.prices-header {
-  color: #666;
-  font-size: 12px;
-}
-#resumo table tbody tr td p.prices {
-  text-align: right;
-  color: #666;
-  font-size: 12px;
-}
-#resumo table tbody tr td p.total-header {
-  font-size: 14px;
-  color: #111;
-}
-#resumo table tbody tr td p.total {
-  font-weight: bold;
-  font-size: 16px;
-  color: #111;
-}
-
-#desconto {
-  font-weight: bold;
-  font-size: 12px;
-  color: #666;
-  padding: 8px 3px !important;
-}
-
-#submit {
-  float: right;
-  margin: 15px 0;
-}
