+++ src/components/checkout/CheckoutEnvio.css
@@ -1,0 +1,98 @@
+/* Compomente Envio */
+#envio {
+  grid-area: envio;
+}
+
+details {
+  border: 1px solid #ccc;
+  border-radius: var(--radius);
+  margin: 15px 0;
+}
+details summary::-webkit-details-marker {
+  display: none;
+}
+details summary {
+  padding: 15px;
+  cursor: pointer;
+  list-style: none;
+  position: relative;
+  width: 100%;
+  background-color: #efefef;
+  border-radius: inherit;
+}
+details summary::after {
+  content: "+";
+  position: absolute;
+  padding-right: 5px;
+  font-size: 2rem;
+  line-height: 0;
+  margin-top: 0.75rem;
+  margin-right: 12px;
+  right: 0;
+  font-weight: 200;
+  transform-origin: center;
+  transition: 200ms linear;
+}
+
+details[open] summary {
+  border-bottom: 1px solid #ccc;
+}
+
+details[open] summary ~ * {
+  animation: open 0.3s ease-in-out;
+}
+
+details[open] summary:after {
+  transform: rotate(45deg);
+  font-size: 2.25rem;
+}
+
+@keyframes open {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+details summary::-webkit-details-marker {
+  display: none;
+}
+#endereco_cont {
+  padding: 15px;
+}
+#endereco_cont:last-child {
+  border-bottom: none;
+}
+#pagamentos {
+  padding: 15px;
+}
+
+.forma_pagamento_titulo {
+  font-size: 13px;
+  font-weight: bold;
+  padding: 10px 0;
+}
+
+.forma_pagamento_desc {
+  font-size: 10px;
+  color: #747474;
+  text-align: right;
+}
+.forma_pagamento_img {
+  width: 20px;
+  height: auto;
+  border-radius: 5px;
+  margin-right: 10px;
+}
+
+#obs_label {
+  margin: 20px 0;
+}
+
+textarea {
+  width: 100%;
+  padding: 15px;
+  border: 1px solid var(--destaque);
+  border-radius: var(--radius);
+}
