+++ src/components/checkout/CheckoutPedido.css
@@ -1,0 +1,123 @@
+/* Compomente Pedido */
+
+#pedido {
+  grid-area: pedido;
+}
+
+#pedido span {
+  display: inline-block;
+}
+
+#pedido span#return {
+  float: right;
+}
+
+#pedido span#return a {
+  font-size: 14px;
+}
+
+#pedido a {
+  font-size: 10px;
+  color: #666;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+#pedido table {
+  width: 98%;
+}
+
+#pedido table td {
+  padding: 12px;
+}
+#pedido table td#preco {
+  width: 120px;
+}
+#pedido table td#empty_list {
+  padding: 0;
+  text-align: center;
+  font-style: italic;
+  line-height: 8.5;
+}
+
+.produto_img {
+  position: relative;
+}
+
+.produto_img img {
+  width: 145px;
+  height: auto;
+}
+
+.produto_nome {
+  font-size: 16px;
+  font-weight: 700;
+}
+
+.produto_ref {
+  font-size: 12px;
+  color: #747474;
+  line-height: 1.85;
+  margin-bottom: 4px;
+}
+
+.produto_atributo {
+  text-align: left;
+  text-decoration: none;
+  font-size: 12px;
+  line-height: 2;
+  /*margin: 8px 0;*/
+}
+
+.delete_produto {
+  text-align: right;
+  margin: 6px;
+}
+
+.produto_preco {
+  text-align: right;
+  font-size: 14px;
+}
+.produto_total {
+  margin-top: 25px;
+  font-weight: bold;
+}
+.dashed {
+  color: #747474;
+  margin-right: 4px;
+  font-size: 12px;
+  text-decoration: line-through;
+}
+.stepper {
+  display: flex;
+  border: 1px solid #e8e8e9;
+  border-radius: var(--radius);
+  background-color: #f8f8f9;
+  height: 32px;
+  width: 88px;
+  margin-top: 6px;
+}
+
+.stepper__button {
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  width: 32px;
+  background: none;
+  border: none;
+  font-size: 2.5em;
+  cursor: pointer;
+}
+
+.stepper__content {
+  flex: 1;
+}
+
+.stepper__value {
+  height: 100%;
+  width: 100%;
+  border: none;
+  background: none;
+  text-align: center;
+  padding-top: 4px;
+}
