+++ src/components/checkout/CheckoutResumo.css
@@ -1,0 +1,47 @@
+/* Compomente Resumo */
+#resumo {
+  grid-area: resumo;
+}
+
+#resumo table {
+  margin: 0 auto;
+  width: 98%;
+  line-height: 1.75;
+}
+
+#resumo table tbody tr .row-div {
+  margin: 7px 0;
+  padding: 3px;
+  border-bottom: 1px solid #efefef;
+}
+
+#resumo table tbody tr td p.prices-header {
+  color: #666;
+  font-size: 12px;
+}
+#resumo table tbody tr td p.prices {
+  text-align: right;
+  color: #666;
+  font-size: 12px;
+}
+#resumo table tbody tr td p.total-header {
+  font-size: 14px;
+  color: #111;
+}
+#resumo table tbody tr td p.total {
+  font-weight: bold;
+  font-size: 16px;
+  color: #111;
+}
+
+#desconto {
+  font-weight: bold;
+  font-size: 12px;
+  color: #666;
+  padding: 8px 3px !important;
+}
+
+#submit {
+  float: right;
+  margin: 15px 0;
+}
